move base_hyperparameter_generator tests to own file
