+++ tests/test_modelgen.py
@@ -51,25 +51,6 @@ class ModelGenerationSuite(unittest.TestCase):
         y_train = np.random.randint(0, 1, size=(1, nr_classes))
         return X_train, y_train
 
-    def test_regularization_is_float(self):
-        """ Regularization should be a float. """
-        reg = modelgen.get_regularization(0, 5)
-        assert type(reg) == np.float
-
-    def test_regularization_0size_interval(self):
-        """ Regularization from zero size interval [2,2] should be 10^-2. """
-        reg = modelgen.get_regularization(2, 2)
-        assert reg == 0.01
-
-    def test_base_hyper_parameters_reg(self):
-        """ Base hyper parameter set should contain regularization. """
-        hyper_parameter_set = modelgen.generate_base_hyper_parameter_set(low_lr=1,
-                                                                         high_lr=4,
-                                                                         low_reg=1,
-                                                                         high_reg=3)
-
-        assert 'regularization_rate' in hyper_parameter_set.keys()
-
     # Tests for CNN model:
     def test_cnn_starts_with_batchnorm(self):
         """ CNN models should always start with a batch normalization layer. """
